refactor: organize Basic Pitch modules
